# 快进快出结果(去除 来源卡号 与 沉淀卡号相同的情况)

See merge request ace/santaizi!86

+++ tldw-analysis-app/src/main/java/com/zqykj/app/service/vo/fund/FastInFastOutResult.java
@@ -43,7 +43,7 @@ public class FastInFastOutResult {
     /**
      * 流入金额
      */
-    private double inflowAmount;
+    private BigDecimal inflowAmount;
 
     /**
      * 资金中转卡号
@@ -68,7 +68,7 @@ public class FastInFastOutResult {
     /**
      * 流出金额
      */
-    private double outflowAmount;
+    private BigDecimal outflowAmount;
 
     /**
      * 资金沉淀卡号
